+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 19. Remove `Version` class (Value Object > Version of software). Use `Version` class from `meritoo/common-library` package.
 20. Base Dependency Injection (DI) extension, the `BaseExtension` class
 21. Add main class of this bundle, the `MeritooCommonBundle` class
+22. Add Dependency Injection (DI) Extension (and configuration) for this bundle
 
 # 0.0.1